Create Sals-shipping
This is one my first Python project, allowing the user to insert the weight of his package and to calculate what's the cheapest shipping method and what would be the total rate:

Sal runs the biggest shipping company in the tri-county area, Sal’s Shippers. In this project, you’ll build a program that will take the weight of a package and determine the cheapest way to ship that package using Sal’s Shippers. Sal’s Shippers has several different options for a customer to ship their package. They have ground shipping, which is a small flat charge plus a rate based on the weight of your package. Premium ground shipping, which is a much higher flat charge, but you aren’t charged for weight. They recently also implemented drone shipping, which has no flat charge, but the rate based on weight is triple the rate of ground shipping.